Move custom JSON encoder/decoders to new file
