+++ fsspec/spec.py
@@ -7,17 +7,14 @@ import os
 import threading
 import warnings
 import weakref
-from contextlib import suppress
 from errno import ESPIPE
 from glob import has_magic
 from hashlib import sha256
-from pathlib import PurePath
-from typing import Any, Callable, ClassVar, Dict, List, Optional, Tuple
+from typing import Any, ClassVar, Dict, Tuple
 
 from .callbacks import DEFAULT_CALLBACK
 from .config import apply_config, conf
 from .dircache import DirCache
-from .registry import _import_class, get_filesystem_class
 from .transaction import Transaction
 from .utils import (
     _unstrip_protocol,
@@ -122,83 +119,6 @@ class AbstractFileSystem(metaclass=_Cached):
     # Set by _Cached metaclass
     storage_args: Tuple[Any, ...]
     storage_options: Dict[str, Any]
-
-    class _JSONEncoder(json.JSONEncoder):
-        def default(self, o: Any) -> Any:
-            if isinstance(o, AbstractFileSystem):
-                return o.to_dict()
-            if isinstance(o, PurePath):
-                cls = type(o)
-                return {"cls": f"{cls.__module__}.{cls.__name__}", "str": str(o)}
-
-            return super().default(o)
-
-    class _JSONDecoder(json.JSONDecoder):
-        def __init__(
-            self,
-            *,
-            object_hook: Optional[Callable[[Dict[str, Any]], Any]] = None,
-            parse_float: Optional[Callable[[str], Any]] = None,
-            parse_int: Optional[Callable[[str], Any]] = None,
-            parse_constant: Optional[Callable[[str], Any]] = None,
-            strict: bool = True,
-            object_pairs_hook: Optional[Callable[[List[Tuple[str, Any]]], Any]] = None,
-        ) -> None:
-            self.original_object_hook = object_hook
-
-            super().__init__(
-                object_hook=self.custom_object_hook,
-                parse_float=parse_float,
-                parse_int=parse_int,
-                parse_constant=parse_constant,
-                strict=strict,
-                object_pairs_hook=object_pairs_hook,
-            )
-
-        @classmethod
-        def try_resolve_path_cls(cls, dct: Dict[str, Any]):
-            with suppress(Exception):
-                fqp = dct["cls"]
-
-                try:
-                    path_cls = _import_class(fqp)
-                except Exception:
-                    raise
-
-                if issubclass(path_cls, PurePath):
-                    return path_cls
-
-            return None
-
-        @classmethod
-        def try_resolve_fs_cls(cls, dct: Dict[str, Any]):
-            with suppress(Exception):
-                fqp = dct["cls"]
-
-                try:
-                    fs_cls = _import_class(fqp)
-                except Exception:
-                    if "protocol" in dct:
-                        return get_filesystem_class(dct["protocol"])
-
-                    raise
-
-                if issubclass(fs_cls, AbstractFileSystem):
-                    return fs_cls
-
-            return None
-
-        def custom_object_hook(self, dct: Dict[str, Any]):
-            if "cls" in dct:
-                if (obj_cls := self.try_resolve_fs_cls(dct)) is not None:
-                    return AbstractFileSystem.from_dict(dct)
-                if (obj_cls := self.try_resolve_path_cls(dct)) is not None:
-                    return obj_cls(dct["str"])
-
-            if self.original_object_hook is not None:
-                return self.original_object_hook(dct)
-
-            return dct
 
     def __init__(self, *args, **storage_options):
         """Create and configure file-system instance
@@ -1477,7 +1397,9 @@ class AbstractFileSystem(metaclass=_Cached):
         multiple), ``args`` (positional args, usually empty), and all other
         keyword arguments as their own keys.
         """
-        return json.dumps(self, cls=self._JSONEncoder)
+        from .json import FilesystemJSONEncoder
+
+        return json.dumps(self, cls=FilesystemJSONEncoder)
 
     @staticmethod
     def from_json(blob: str) -> AbstractFileSystem:
@@ -1499,7 +1421,9 @@ class AbstractFileSystem(metaclass=_Cached):
         This can import arbitrary modules (as determined by the ``cls`` key)
         at runtime.
         """
-        return json.loads(blob, cls=AbstractFileSystem._JSONDecoder)
+        from .json import FilesystemJSONDecoder
+
+        return json.loads(blob, cls=FilesystemJSONDecoder)
 
     def to_dict(self) -> Dict[str, Any]:
         """
@@ -1542,7 +1466,9 @@ class AbstractFileSystem(metaclass=_Cached):
         This can import arbitrary modules (as determined by the ``cls`` key)
         at runtime.
         """
-        cls = AbstractFileSystem._JSONDecoder.try_resolve_fs_cls(dct)
+        from .json import FilesystemJSONDecoder
+
+        cls = FilesystemJSONDecoder.try_resolve_fs_cls(dct)
         if cls is None:
             raise ValueError("Not a serialized AbstractFileSystem")
